third_party/WebKit/LayoutTests/fast/dom/shadow/content-combinator-css-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style id="style1">
.foo /content/ div .bar::before { display: block; }
div content /content/ div content /content/ div.green { color: green; }
* /content/ * { color: blue; }
</style>
</head>
<script>
description("Test for cssText of '::distributed' rule.");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(0).cssText", ".foo /content/ div .bar::before { display: block; }");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(1).cssText", "div content /content/ div content /content/ div.green { color: green; }");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(2).cssText", "* /content/ * { color: blue; }");
</script>
</html>
